# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Memory Threads API - AI-powered memory engine for files
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /timeline:
    get:
      operationId: getTimeline
      tags: [timeline]
      summary: Get chronological timeline of memories and events
      parameters:
        - in: query
          name: instanceId
          schema:
            type: integer
        - in: query
          name: dateFrom
          schema:
            type: string
        - in: query
          name: dateTo
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
  /timeline/events:
    post:
      operationId: createTimelineEvent
      tags: [timeline]
      summary: Create a custom timeline event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimelineEventRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineEvent"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /timeline/events/{id}:
    put:
      operationId: updateTimelineEvent
      tags: [timeline]
      summary: Update or override a timeline event (works for auto-generated and custom events)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimelineEventRequest"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineEvent"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTimelineEvent
      tags: [timeline]
      summary: Hide an auto-generated event or delete a custom event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
components:
  schemas:
...
    TimelineEvent:
      type: object
      required:
        - id
        - date
        - title
        - type
      properties:
        id:
          type: string
        date:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [upload, event, deadline, mention]
        memoryId:
          type: ["integer", "null"]
        memoryName:
          type: ["string", "null"]
        isCustom:
          type: boolean
        isEdited:
          type: boolean
    CreateTimelineEventRequest:
      type: object
      required:
        - date
        - title
      properties:
        date:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [upload, event, deadline, mention]
        memoryId:
          type: ["integer", "null"]
        instanceId:
          type: integer
    UpdateTimelineEventRequest:
      type: object
      properties:
        date:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [upload, event, deadline, mention]
...
    DeleteResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string